server: Move api call processing into `WSChiaConnection`
