+++ chia/server/ws_connection.py
@@ -6,7 +6,8 @@ import logging
 import time
 import traceback
 from dataclasses import dataclass, field
-from typing import Any, Dict, List, Optional, Tuple, Union
+from secrets import token_bytes
+from typing import Any, Awaitable, Callable, Dict, List, Optional, Set, Tuple, Union
 
 from aiohttp import ClientSession, WSCloseCode, WSMessage, WSMsgType
 from aiohttp.client import ClientWebSocketResponse
@@ -15,8 +16,8 @@ from typing_extensions import Protocol, final
 
 from chia.cmds.init_funcs import chia_full_version_str
 from chia.protocols.protocol_message_types import ProtocolMessageTypes
-from chia.protocols.protocol_state_machine import message_response_ok
-from chia.protocols.protocol_timing import INTERNAL_PROTOCOL_ERROR_BAN_SECONDS
+from chia.protocols.protocol_state_machine import message_requires_reply, message_response_ok
+from chia.protocols.protocol_timing import API_EXCEPTION_BAN_SECONDS, INTERNAL_PROTOCOL_ERROR_BAN_SECONDS
 from chia.protocols.shared_protocol import Capability, Handshake
 from chia.server.outbound_message import Message, NodeType, make_msg
 from chia.server.rate_limits import RateLimiter
@@ -35,6 +36,7 @@ from chia.util.streamable import Streamable
 LENGTH_BYTES: int = 4
 
 WebSocket = Union[WebSocketResponse, ClientWebSocketResponse]
+ConnectionCallback = Callable[["WSChiaConnection"], Awaitable[None]]
 
 
 class ConnectionClosedCallbackProtocol(Protocol):
@@ -57,6 +59,7 @@ class WSChiaConnection:
     """
 
     ws: WebSocket
+    api: Any
     local_type: NodeType
     local_port: int
     local_capabilities_for_handshake: List[Tuple[uint16, str]]
@@ -74,8 +77,12 @@ class WSChiaConnection:
     is_outbound: bool
 
     # Messaging
-    incoming_queue: asyncio.Queue[Tuple[Message, WSChiaConnection]]
+    received_message_callback: Optional[ConnectionCallback]
+    incoming_queue: asyncio.Queue[Message] = field(default_factory=asyncio.Queue)
     outgoing_queue: asyncio.Queue[Message] = field(default_factory=asyncio.Queue)
+    api_tasks: Dict[bytes32, asyncio.Task[None]] = field(default_factory=dict)
+    # Contains task ids of api tasks which should not be canceled
+    execute_tasks: Set[bytes32] = field(default_factory=set)
 
     # ChiaConnection metrics
     creation_time: float = field(default_factory=time.time)
@@ -85,6 +92,7 @@ class WSChiaConnection:
 
     peer_server_port: Optional[uint16] = None
     inbound_task: Optional[asyncio.Task[None]] = None
+    incoming_message_task: Optional[asyncio.Task[None]] = None
     outbound_task: Optional[asyncio.Task[None]] = None
     active: bool = False  # once handshake is successful this will be changed to True
     _close_event: asyncio.Event = field(default_factory=asyncio.Event)
@@ -105,11 +113,12 @@ class WSChiaConnection:
         cls,
         local_type: NodeType,
         ws: WebSocket,
+        api: Any,
         server_port: int,
         log: logging.Logger,
         is_outbound: bool,
+        received_message_callback: Optional[ConnectionCallback],
         peer_host: str,
-        incoming_queue: asyncio.Queue[Tuple[Message, WSChiaConnection]],
         close_callback: Optional[ConnectionClosedCallbackProtocol],
         peer_id: bytes32,
         inbound_rate_limit_percent: int,
@@ -133,6 +142,7 @@ class WSChiaConnection:
 
         return cls(
             ws=ws,
+            api=api,
             local_type=local_type,
             local_port=server_port,
             local_capabilities_for_handshake=local_capabilities_for_handshake,
@@ -146,7 +156,7 @@ class WSChiaConnection:
             outbound_rate_limiter=RateLimiter(incoming=False, percentage_of_limit=outbound_rate_limit_percent),
             inbound_rate_limiter=RateLimiter(incoming=True, percentage_of_limit=inbound_rate_limit_percent),
             is_outbound=is_outbound,
-            incoming_queue=incoming_queue,
+            received_message_callback=received_message_callback,
             session=session,
         )
 
@@ -226,6 +236,7 @@ class WSChiaConnection:
 
         self.outbound_task = asyncio.create_task(self.outbound_handler())
         self.inbound_task = asyncio.create_task(self.inbound_handler())
+        self.incoming_message_task = asyncio.create_task(self.incoming_message_handler())
 
     async def close(
         self,
@@ -255,6 +266,8 @@ class WSChiaConnection:
         try:
             if self.inbound_task is not None:
                 self.inbound_task.cancel()
+            if self.incoming_message_task is not None:
+                self.incoming_message_task.cancel()
             if self.outbound_task is not None:
                 self.outbound_task.cancel()
             if self.ws is not None and self.ws.closed is False:
@@ -262,6 +275,7 @@ class WSChiaConnection:
             if self.session is not None:
                 await self.session.close()
             self.cancel_pending_requests()
+            self.cancel_tasks()
         except Exception:
             error_stack = traceback.format_exc()
             self.log.warning(f"Exception closing socket: {error_stack}")
@@ -288,6 +302,12 @@ class WSChiaConnection:
             except Exception as e:
                 self.log.error(f"Failed setting event for {message_id}: {e} {traceback.format_exc()}")
 
+    def cancel_tasks(self) -> None:
+        for task_id, task in self.api_tasks.copy().items():
+            if task_id in self.execute_tasks:
+                continue
+            task.cancel()
+
     async def outbound_handler(self) -> None:
         try:
             while not self.closed:
@@ -311,6 +331,96 @@ class WSChiaConnection:
                 self.log.error(f"Exception: {e} with {self.peer_host}")
                 self.log.error(f"Exception Stack: {error_stack}")
 
+    async def _api_call(self, full_message: Message, task_id: bytes32) -> None:
+        start_time = time.time()
+        message_type = ""
+        try:
+            if self.received_message_callback is not None:
+                await self.received_message_callback(self)
+            self.log.debug(
+                f"<- {ProtocolMessageTypes(full_message.type).name} from peer " f"{self.peer_node_id} {self.peer_host}"
+            )
+            message_type = ProtocolMessageTypes(full_message.type).name
+
+            f = getattr(self.api, message_type, None)
+
+            if f is None:
+                self.log.error(f"Non existing function: {message_type}")
+                raise ProtocolError(Err.INVALID_PROTOCOL_MESSAGE, [message_type])
+
+            metadata = get_metadata(function=f)
+            if metadata is None:
+                self.log.error(f"Peer trying to call non api function {message_type}")
+                raise ProtocolError(Err.INVALID_PROTOCOL_MESSAGE, [message_type])
+
+            # If api is not ready ignore the request
+            if hasattr(self.api, "api_ready"):
+                if self.api.api_ready is False:
+                    return None
+
+            timeout: Optional[int] = 600
+            if metadata.execute_task:
+                # Don't timeout on methods with execute_task decorator, these need to run fully
+                self.execute_tasks.add(task_id)
+                timeout = None
+
+            if metadata.peer_required:
+                coroutine = f(full_message.data, self)
+            else:
+                coroutine = f(full_message.data)
+
+            async def wrapped_coroutine() -> Optional[Message]:
+                try:
+                    # hinting Message here is compensating for difficulty around hinting of the callbacks
+                    result: Message = await coroutine
+                    return result
+                except asyncio.CancelledError:
+                    pass
+                except Exception as e:
+                    tb = traceback.format_exc()
+                    self.log.error(f"Exception: {e}, {self.get_peer_logging()}. {tb}")
+                    raise
+                return None
+
+            response: Optional[Message] = await asyncio.wait_for(wrapped_coroutine(), timeout=timeout)
+            self.log.debug(
+                f"Time taken to process {message_type} from {self.peer_node_id} is "
+                f"{time.time() - start_time} seconds"
+            )
+
+            if response is not None:
+                response_message = Message(response.type, full_message.id, response.data)
+                await self.send_message(response_message)
+            # check that this call needs a reply
+            elif message_requires_reply(ProtocolMessageTypes(full_message.type)) and self.has_capability(
+                Capability.NONE_RESPONSE
+            ):
+                # this peer can accept None reply's, send empty msg back, so it doesn't wait for timeout
+                response_message = Message(uint8(ProtocolMessageTypes.none_response.value), full_message.id, b"")
+                await self.send_message(response_message)
+        except TimeoutError:
+            self.log.error(f"Timeout error for: {message_type}")
+        except Exception as e:
+            if not self.closed:
+                tb = traceback.format_exc()
+                self.log.error(f"Exception: {e} {type(e)}, closing connection {self.get_peer_logging()}. {tb}")
+            else:
+                self.log.debug(f"Exception: {e} while closing connection")
+            # TODO: actually throw one of the errors from errors.py and pass this to close
+            await self.close(API_EXCEPTION_BAN_SECONDS, WSCloseCode.PROTOCOL_ERROR, Err.UNKNOWN)
+        finally:
+            if task_id in self.api_tasks:
+                self.api_tasks.pop(task_id)
+            if task_id in self.execute_tasks:
+                self.execute_tasks.remove(task_id)
+
+    async def incoming_message_handler(self) -> None:
+        while True:
+            message = await self.incoming_queue.get()
+            task_id: bytes32 = bytes32(token_bytes(32))
+            api_task = asyncio.create_task(self._api_call(message, task_id))
+            self.api_tasks[task_id] = api_task
+
     async def inbound_handler(self) -> None:
         try:
             while not self.closed:
@@ -321,7 +431,7 @@ class WSChiaConnection:
                         event = self.pending_requests[message.id]
                         event.set()
                     else:
-                        await self.incoming_queue.put((message, self))
+                        await self.incoming_queue.put(message)
                 else:
                     continue
         except asyncio.CancelledError:
