+++ chia/simulator/time_out_assert.py
@@ -48,7 +48,7 @@ def time_out_messages(incoming_queue: asyncio.Queue, msg_name: str, count: int =
         if incoming_queue.qsize() < count:
             return False
         for _ in range(count):
-            response = (await incoming_queue.get())[0].type
+            response = (await incoming_queue.get()).type
             if ProtocolMessageTypes(response).name != msg_name:
                 # log.warning(f"time_out_message: found {response} instead of {msg_name}")
                 return False
