+++ tests/connection_utils.py
@@ -41,7 +41,6 @@ async def add_dummy_connection(
 ) -> Tuple[asyncio.Queue, bytes32]:
     timeout = aiohttp.ClientTimeout(total=10)
     session = aiohttp.ClientSession(timeout=timeout)
-    incoming_queue: asyncio.Queue = asyncio.Queue()
     config = load_config(server.root_path, "config.yaml")
     chia_ca_crt_path, chia_ca_key_path = chia_ssl_ca_paths(server.root_path, config)
     dummy_crt_path = server.root_path / "dummy.crt"
@@ -58,11 +57,12 @@ async def add_dummy_connection(
     wsc = WSChiaConnection.create(
         type,
         ws,
+        server.api,
         server._port,
         log,
         True,
+        server.received_message_callback,
         self_hostname,
-        incoming_queue,
         None,
         peer_id,
         100,
@@ -70,7 +70,9 @@ async def add_dummy_connection(
         local_capabilities_for_handshake=capabilities,
     )
     await wsc.perform_handshake(server._network_id, protocol_version, dummy_port, NodeType.FULL_NODE)
-    return incoming_queue, peer_id
+    if wsc.incoming_message_task is not None:
+        wsc.incoming_message_task.cancel()
+    return wsc.incoming_queue, peer_id
 
 
 async def connect_and_get_peer(server_1: ChiaServer, server_2: ChiaServer, self_hostname: str) -> WSChiaConnection:
