+++ tests/core/ssl/test_ssl.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import asyncio
-
 import aiohttp
 import pytest
 
@@ -20,17 +18,17 @@ async def establish_connection(server: ChiaServer, self_hostname: str, ssl_conte
     timeout = aiohttp.ClientTimeout(total=10)
     dummy_port = 5  # this does not matter
     async with aiohttp.ClientSession(timeout=timeout) as session:
-        incoming_queue: asyncio.Queue = asyncio.Queue()
         url = f"wss://{self_hostname}:{server._port}/ws"
         ws = await session.ws_connect(url, autoclose=False, autoping=True, ssl=ssl_context)
         wsc = WSChiaConnection.create(
             NodeType.FULL_NODE,
             ws,
+            server.api,
             server._port,
             server.log,
             True,
+            server.received_message_callback,
             self_hostname,
-            incoming_queue,
             None,
             bytes32(b"\x00" * 32),
             100,
